requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
debugpy
decorator
exceptiongroup
executing
Flask==3.0.2
fonttools==4.53.0
frozendict==2.4.4
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
html5lib==1.1
httplib2==0.22.0
idna==3.7
importlib_metadata
ipykernel
ipython
ipywidgets==8.1.3
joblib==1.4.2
jupyter_client
jupyter_core
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.0
matplotlib-inline
multitasking==0.0.11
nest_asyncio
newsapi-python==0.2.7
numpy==2.0.0
packaging
pandas==2.2.2
parso
peewee==3.17.5
pickleshare
pillow==10.3.0
platformdirs
prompt_toolkit
proto-plus==1.24.0
protobuf==4.25.3
psutil
pure-eval
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments
pyparsing==3.1.2
python-dateutil
pytz==2024.1
pyzmq
requests==2.32.3
rsa==4.9
scikit-learn==1.5.0
scipy==1.14.0
setuptools==70.1.1
six
soupsieve==2.5
stack-data
tensorflow==2.19.0
threadpoolctl==3.5.0
tornado
tqdm==4.66.4
traitlets
typing_extensions
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth
webencodings==0.5.1
wheel==0.43.0
widgetsnbextension==4.0.11
yfinance==0.2.55
zipp
